refactored, added deprecation warnings and rewrote mlflow client
