+++ splicemachine.egg-info/requires.txt
@@ -1,0 +1,13 @@
+atomicwrites==1.1.5
+attrs==18.1.0
+more-itertools==4.2.0
+pluggy==0.6.0
+py==1.5.3
+py4j==0.10.7
+pytest==3.6.1
+six==1.11.0
+mlflow==0.4.2
+graphviz==0.8.4
+numpy==1.15.0
+h2o_pysparkling_2.2
+pandas==0.22.0
